refactor(importers): flatten src/importers/aos4 into src/importers
`aos4` was the only directory under `src/importers/`, so the qualifier
distinguished nothing, and it read confusingly against `src/aos4/import/` --
two import trees whose names differ only in word order.

The split between the two is architecturally forced and stays: `src/aos4/` may
not depend outward, so provider-facing parsing (official app text, Listbot,
New Recruit `.ros`/`.rosz`/`.json`) lives outside the domain while roster
resolution stays inside it. Only the redundant path segment goes away.

Relative imports into the domain shorten by one level, from `../../aos4/import`
to `../aos4/import`. Sibling imports are untouched. The AGENTS.md architecture
notes and the two import fixture READMEs are updated to the new paths.

No behavior changes; the edits are import paths only.
