+++ src/importers/listbot.ts
@@ -1,8 +1,4 @@
-import type {
-  Aos4ParsedRosterResult,
-  ParsedRosterSelection,
-  ParsedRosterSelectionKind,
-} from '../../aos4/import'
+import type { Aos4ParsedRosterResult, ParsedRosterSelection, ParsedRosterSelectionKind } from '../aos4/import'
 import type { Aos4ImportLine } from './detectTextSource'
 
 const footerPattern = /^Generated by Listbot 4\.0$/
