+++ src/importers/rosterJson.ts
@@ -1,4 +1,4 @@
-import type { Aos4ParsedRosterResult } from '../../aos4/import'
+import type { Aos4ParsedRosterResult } from '../aos4/import'
 import { parseAos4RosterTree, rosterError, type RosterNode } from './rosterTree'
 
 /**
